requirements.txt
цasciimatics==1.11.0
astroid==2.3.3
autopep8==1.5.2
configparser==5.0.0
confuse==1.1.0
entrypoints==0.3
flake8==3.7.9
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
Pillow==7.1.1
pycodestyle==2.5.0
pyfiglet==0.8.post1
pyflakes==2.1.1
pygame==1.9.3
pylint==2.4.4
PyTMX==3.21.5
PyYAML==5.3.1
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wcwidth==0.1.9
wrapt==1.11.2
